Set correct shell for root
On EC2 and Azure, root user cannot have a password set. When it was
changed in the past, it was done in a way root shell changed from
/bin/sh to /etc/rc.initial. This change had a side effect on some
unbound operations that are executed using 'su -m unbound'.

Fix it skipping only password set for root user but making sure shell
is changed to /bin/sh.