Fix #6869 diag_routes resolve names for RELENG_2_3
This code to parse the netstat output and use gethostbyaddr() to reverse resolve names is only needed in RELENG_2_3, so that long names are not truncated. In FreeBSD 10.3 some long resolved names are still being truncated by `netstat`, even though `-W` is specified.

In FreeBSD 11 for pfSense 2.4, `netstat -rW` correctly outputs the full resolved name, even when it is very long. So that all passes through to the Diag Routes GUI display correctly.