DHCPv6 server range from-to check
When track6 is on, the particular IPv6 address of the interface is not known (and could change from time to time). So $ifcfgip is set to "::", and $ifcfgsn to 64.
Unfortunately gen_subnetv6("::", 64) returns blank "". That no longer passes the checks done by recent changes to is_inrange_v6().
To fix this we can just explicitly set
```
$subnet_start = "0::";
```
for this case.

Alternatively we could add code to gen_subnetv6() to handle the special case of being passed "::" as the IP address.
Actually the underlying issue is that the Net_IPv6 compress() function seems to be compressing "0:0:0:0:0:0:0:0" to "", when I would hope it would compress to "0::" or "::0". But that requires some upstream changes, and sorting out which of "0::" and "::0" is the correct representation.