Save the product name on /etc/product_name at /etc/rc.banner.
This way we do not need to run PHP again on rc.initial to read the product name.

This file is only written if /etc is writeable, if /etc/product_name does not exist the old (and slow) PHP code will run.

This make the initial menu a bit faster on uFW.