Feature #3151 Disable gateway monitoring actions
without disabling gateway monitoring.

This allows the user to continue to monitor the gateway with dpinger, so
they can see how it is performing, but for the system not to take any
real action if the latency/loss exceeds the given limits.

A typical use case for this would be on a single-WAN system. There is no
failover option, so there is no point taking any real action when the
latency/loss is high. Having stuff try to failover (and stop/start
stuff...) is just disruptive.

In ths case the use could have disabled monitoring completely, but then
they get no feedback abut gateway performance.