DHCP server edit when DHCP Relay enabled
Part of Redmine #6997
1) Display the DHCP Server settings even when DHCP Relay is enabled, but
disable the "enable" checkbox in this case - so the user cannot enable
DHCP Server when DHCP Relay is enabled, but can view and modify the DHCP
Server parameters (ready for the day they want to use it).
2) Actually validate the POSTed DHCP Server parameters, even if it is
disabled. Previously you could put rubbish values in fields and they
would be "happily" saved when DHCP Server was disabled. We want users to
be able to setup DHCP Server ready-to-enable and knowing that they have
put validated stuff in.
3) When DHCP Server is disabled, allow Range From and Range To to be
left empty. When first setting up parameters (e.g. with the UI page up
for the first time) it is OK for the user to be able to just press
"Save". e.g. they are just pressing Save as a way to get out (as if it
is a Cancel button). We do not need to force them to enter range from/to
IP addresses.