dhcp6c no release option
Some ISP's will release the allocated PD or address on receipt of the
release signal and will then issue a new PD or address on next request.
This is not always wanted. dhcp6c sends this signal by default. A
modification to dhcp6c has the PR #229 in pfsense/FreeBSD-ports. The -n
option is added to the start up of dhcp6c if this option is set.