User Define DUID
Moved the Configuration of the DUID to System->Advanced->Networking.

DUID validity is checked for content and length. then creates a system
config variable called global-v6duid.
Can replace PR #3294

In Interfaces WAN an option exists to use the global DUID ( which has
been defined or it flags an error ). On WAN interface configure the
write of the DUID file takes place by calling the function in util.inc