validateipaddr return address family
Enhanced the return values from validateipaddr() so the caller can know if the validated address is IPv4 or IPv6 (or an alias if that check was requested). That makes it a better replacement for is_ipaddr().
(cherry picked from commit 24eb39e29641488c7fa970d69f42edd84e2fbb1e)