Add a factory post installation handler.
It is triggered by the presence of the file '/root/factory_boot' and once read, the file is deleted (runs one time only).

When enabled, it adds a 5 seconds timeout on initial menu that halts the system when the timeout timer triggers.

This is intended to automate the factory installation process.