improve CSS handling for icmp types (overflow/table)
Minor CSS and formatting improvement to layout for icmp types in rules table. With this change, if several icmp types are specified in the rule, if they won't fit into 2 lines the GUI automatically switches to a scrollable overflow instead, to preserve the table layout. Also underlining is per icmptype so commas etc aren't underlined which is much easier to read