Alternate fix for jim-p's point that seems to resolve everything neatly
@jim-p commented on the PR that:

> This change is unwarranted. The protocol default should remain TCP, it is set that way on purpose (otherwise people get confused by the lack of port options being visible). It's also not relevant to the other changes being made on this PR.

An alternate fix for jim-p's point. this only affects creating new rules and I take the point. I found this a much better fix (AFAICS).  It seems to resolve all issues neatly.

Proposed fix:  leave "any" at the top of the list as that's the logical place people almost always look for it if they want it, _but set the default proto to tcp for new rules_ so that ports and other expected items are displayed by default too. After all, the default protocol is only relevant for showing tcp and ports fields, when a new blank rule is created (obvious: if the rule exists it would display the protocol in the existing rule). 

@jim-p can you try this as a fix, and see if it would be acceptable for resolving your point?