Check for static routes wholly within interface subnet
The existing code checks for an interface subnet and static route subnet
that are both exactly the same, and puts up an input error for that.
Actually if the static route subnet is contained in the interface
subnet, then we should put up an input error - because that is pretty
dumb to try and static route some part of a local interafce subnet away
out some gateway. Probably it won't break things (the routing will
probably deliver the local packets direct to the local interface anyway,
effectively ignoring the static route). But it seems good to let the
user know that what they are doing is nonsense.
Or is there some use case that I ave not thought about?

Note: If the static route covers a range bigger than an interface
subnet, that is OK. Packets for addresses within the interface are
delivered locally, packets for the rest of the static routed network
will be routed to the specified gateway.