Fix #3454 Do not show Mark All as Read button when no priv
If the user does not have access to index.php then the "Mark All as Read" button for the notices popup does not work for them anyway, so do not show it.
This fixes the obvious UI inconsistency - where the user has a button that they press, but it is not effective.
(cherry picked from commit fe80b3aac6ddd661c7a2daf52ad54f1722915590)