Report problems applying changes
At present various code keeps track of the return value of functions
that are called to apply changes ($retval). Then calls:
```get_std_save_message($retval)```
But get_std_save_message() always returns the same text "The changes
have been applied successfully."

So why is there are parameter to get_std_save_message()?

If we fix this up, returning different text to indicate that there is a
problem, then the subsequent print_info_box() that actually puts up the
message needs to pass the relevant severity also (success, warning,...).
So to minimize code that needs to be changed/duplicated in many places,
I have made print_apply_result_box() which takes $retval and sorts out
both the message text and severity all in 1 go.

system_routes has the typical changes that would be needed in the
top-level code to implement this.

Without changing top-level code, the changes in guiconfig.inc would
result in the problem text being reported when there is a problem, but
in the "success" color.