Captive portal: rework logging and RADIUS accounting when disabling a zone or rebooting
Make captiveportal_radius_stop_all() log the disconnections in the system log and fix it so that it works with the zone id parameter and sends complete RADIUS accounting packets.

Since several zones can share the same RADIUS server, send an Accounting-Off packet only when rebooting, not when disabling a zone.