Implement a different method to obtain the 'main' IP (v4) for the interface.
The last interface IP is always saved in /var/db/${interface}_ip.  Use that file, if it exist, to find the main interface IP.

The file is created by the same process and function that call 'ifconfig setfirst', so the presence of that file should produce a very similar behavior.

If the file does not exist, fallback to previous behavior (return the first IPv4 found on interface).

Ticket #7042