Cache authentication for a short time
Currently there is an authentication lookup for every GET; a single page load can cause dozens of lookups and hundreds of lines of logs (and continuous logs in the case of pages with widgets). This change allows pfsense to remember authentication, but forces recheck after a short time in case of access revocations/changes.