Make appropriate success or danger boxes in system_crlmanager
When there is a problem, the messages should be in a "danger" box.
Note: It is a bit difficult to actual get the danger messages to appear in the UI, because the relevant "delete" buttons are not shown in the UI when delete is not valid. But it can be tested by starting to edit a CRL that is in use, then change the "edit=" to "del=" in the URL bar and send it.
(cherry picked from commit 61d1ca3e0ffb0fb6a397edb8ca5d88f039787140)