Correctly report unmonitored gateway status
If an alternate monitor IP has been entered and saved, then the user
checks "Disable Gateway Monitoring" and saves, the alternate monitor IP
is retained in the config - that is handy for when unchecking "Disable
Gateway Monitoring" later on.
But the Gateways widget and Status Gateways do not correctly understand
this combination. The gateway status shows as "Online" when it is
intended to show "Online (unmonitored)".

This PR corrects this.

(cherry picked from commit 0c5d4e8d3e4dc81a9c7eb883a40296493e9faa2b)