Allow hiding of gateways in Gateways Widget
Sometimes there are gateways that I do not care about, so it is handy to
be able to save some space on the dashboard by hiding them.
This uses similar code structure taken from service_status widget. I
made the Save and Clear buttons and all that stuff use the same styles
also. I guess that the way it was all done in the Services widget is the
more "modern" way.