clarify type of firewall rule needed
A comment on the original PR says that it's not clear if the rule for this needs to be IPv6 or IPv4 with protocol IPv6, and users get confused. I've tried to clarify it here.  I've also tried to "de-mystify" it, because it's not that tunneling needs special rules, it's just that any traffic needs rules and tunneled traffic is handled as IPv6 for rules purposes, so it needs an IPv6 rule. But the user may want other rules, or selective rules, so I'[ve tried to be clear this isn't about some "special" IPv6 rule, but rather, just that any traffic desired to pass, will need one or more IPv6 rules permitting it.
(cherry picked from commit 3509ad4f9215ff46d7a8684e0559bb93c1bd7a78)