Crash Reporter report when no data found
The glob for $crash_files always returns an array - an empty one in the case when there are no matching crash files. So the echo at old line 115 never happened. In any case, if the echo happens, it is spat out as plain text on the GUI outside of all the nice boxes and menus.

Might be nice for the devs to get explicit feedback about the "missing" parts of a crash report - that there were either "No PHP errors found" or "No FreeBSD crash data found".
(cherry picked from commit cb6b7189bed635976ff33377e68f54b36b2647c9)