Make in-memory DNS Server list zero-based
The backend code that handles repeatable field groups and the adding and deleting of rows expects that the field names are zero-based. So make the form have fields dns0, dns1, dns2 etc. These end up in a generic array in the config anyhow, so no hassle here.
For the DNS gateways, it is a bit more tricky. Currently they are named keys in the config 'dns1gw', 'dns2gw, 'dns3gw'...
When being processed by the backend code that handles repeatable field groups, the names need to be zero-based and the number needs to be at the end of the name. So 'dns1gw' in the config becomes 'dnsgw0' in $pconfig and in $_POST. Then the $_POSTed 'dns0' and 'dnsgw0' are validated. Then 'dns0' will end up at the start of the 'dnsservers' array, and 'dnsgw0' will be put in the config using key 'dns1gw'.
This processes the data in the GUI using the repeatable rows code, while preserving the existing method of storing the data in the config.