Add username to logout menu and icon
I find it difficult to work out who I am logged in as. Other software often provides some indication of who is the current logged-in user. An easy way to do this is to add it to the "Log out" text. Then it can be seen either by hovering over the "Log out" icon, or in the System menu.

I added it to both places. But maybe it should just be associated with the icon, and not the System->Logout menu entry?

I also notice that the System menu has "Logout" (all one word), and the hover text for the icon is "Log out" (2 words). These should be consistent - but which one is preferred?