RTSOLD lock creation, dhcp6c launch & kill changes,
Added lock file creation and check to RTSOLD script creation. This is to
prevent mutliple launches of dhcp6c this appears to happen when multiple
RA's are received in rapid succession at the start of a session. Once
created dhcp6c cannot be launched again until the lock file is deleted,
this is done within the kill_dhcp6_client process locking the two
together.

The kill -9 used to kill the dhcp6c client is changed to -15, -9 causes
the process to exit without sending a release if required, and if the
timing is just rignt can cause the pid file to be left behind; -15
allows for a graceful exit and if the release flag is not set then it
sends and waits for the release confirnation.

The launch of dhcp6c when in dhcp6withoutRA is moved to its own
function, as much as anything this makes the code tidy around the bottom
of interface_dhcpv6_configure().

These changes require the use of a new dhcp6c client, the PR for this is
#258 and this has also been PR'd upstream. The new client contains an
extra flag '-x' . This flag is used to prevent the exextion of the same
script by dhcp6c and RTOSOLD at the same time, the flag causes dhcp6c to
not execute the script on receipt of a reply to a request. In
dhcp6cwithoutRA mode then RTSOLD will execute the script after dhcp6c
has been launched and when the RA has been received. This should also
prevent a problem where some ISP's BNG's are tardy when it comes to
replying to the RS.

This has been tested to death on ISP's that require dhcp6withoutRA and I
have tested it using a second pfSense unit as a sub off my main..