Fix #7134 Interface Stats widget ifdescrs not defined
When doing testing I had the Interfaces widget open as well as Interface Statistics. $ifdescrs was already populated as an array by the Interfaces widget, so Interface Statistics had (accidentally) gained the benefit.
If you have just the Interface Statistics widget on the dashboard then you get the reported PHP error, in the code that sets up the checkbox array of selectable interfaces. The problem is fixed by always calling get_configured_interface_with_descr() at the top of this code, rather than calls that were just inside various if() cases.
Note: get_configured_interface_with_descr() in util.inc always returns an array, even if an empty one, so the ```if (!is_array($ifdescrs))``` testing should not be needed.