RTSOLD lock creation, dhcp6c launch & kill changes #2
Added lock file creation and check to RTSOLD script creation. This is to
prevent mutliple launches of dhcp6c, this appears to happen when
multiple RA's are received in rapid succession at the start of a
session. Once created dhcp6c cannot be launched again until the lock
file is deleted, this is done within the kill_dhcp6_client process
locking the two together.

The kill -9 used to kill the dhcp6c client is
changed to -15, -9 causes the process to exit without sending a release
if required, and if the timing is just rignt can cause the pid file to
be left behind; -15 allows for a graceful exit and if the release flag
is not set then it sends and waits for the release confirnation.

The
launch of dhcp6c when in dhcp6withoutRA is moved to its own function, as
much as anything this makes the code tidy around the bottom of
interface_dhcpv6_configure().

A completely new method of implimenting dhcp6wihtoutRA is used. In
default mode RTSOLD launches dhcp6c.  In dhcp6wihtoutRA mode dhcp6
launches RTSOLD.  New scripts are created and old ones modified to
handle this mode, the dhcp6 conf file changes depending on the mode
calling a different script for each mode. In simple terms its
dhcp6->rtsold->wan_configure. Whenever dhcp6 gets a response that
launches its script then it will run rtsold, the RA in turn will cause
the wan6 configure script to run. This method also means the script only
ever runs once and no modified dhcp6c is required.