Make Backup & Restore buttons more robust
I had set the UI to Turkish and then tried to 'Download configuration as XML' and it did not work. 'Download configuration as XML' is not yet translated in Turkish, but the single word 'Download' is translated. So the assumption in the code that the translation of 'Download' is contained somewhere in the translation of 'Download configuration as XML' is not true.

These should use matches of the translations of the same string as used in the button.