Okay, I think this fixes historical icmptype handling
in most cases, if icmptype data for an existing/associated rule is missing or invalid, and we're editing the rule, we want to treat it as "all icmptypes", based on the relevant ipprotocol. But if ipprotocol isn't valid or defined for the existing/associated rule, we can't do that, because we don't know what icmptypes to substitute, because that depends on the ip protocol of the rule. So implode() was failing because the 2nd arg which should be an array, was undefined.

 (I still don't understand how that situation can happen in an existing ruleset, but let's assume it can based on feedback above, no other reason seems to make sense for why implode could have an invalid 2nd argument).

So this fixes it by making no assumption about the ip protocol. We often need to unset icmptype if it's not valid or determinable, so I've changed the previous fix, it now takes the output variable by reference which allows it to unset() it as well as assign a value, and an optional 4th arg which probably won't be needed but ensures fixability, so the calling code can force a value for ipprotocol rather than just unset the result, if necessary.

Probably overdone, and perhaps in the end, would have been simpler to fix via a config update, but should do the job for existing configs and rules?