Update version string at end of boot RELENG_2_3
When there is an upgrade, the echo here was outputting a stale value of the version. For example, on first upgrade from 2.3.3-DEVELOPMENT to 2.3.3-RC the console had:

pfSense (pfSense) 2.3.3-DEVELOPMENT amd6 Sat Feb 11 14:24:27 CST 2017
Bootup complete

FreeBSD/amd64 (myhost.localdomain) (ttyv0)

*** Welcome to pfSense 2.3.3-RC (amd64 full-install) on myhost ***

That is a bit confusing for users to be sure which version it is at this point.
(cherry picked from commit c0044174800c3b509dd79906aeaf69e4c6ff1385)