Use cached groups in get_user_privileges
This should help with https://redmine.pfsense.org/issues/7253
Calls to userHasPrivilege(), which calls get_user_privileges() will be able to complete out of the cache. So this should be no more of a pain when the LDAP or RADIUS server is not responding than existing calls to getAllowedPages().
This code to utilitize a cache of allowed groups is taken from the existing code in getAllowedPages() and massaged a bit to use the way $user is passed in here.